The Gun and the Bullet
A colored background with size was added, as well a gun on the bottom of the screen. Added movement left and right for the gun. Pressing 'a' moves it to the left, 'd' to the right. I also made it possible to move the gun past the side boundaries of the window to get over to the other side. Shooting a bullet is also possible, but the bullet "resets" itself every time ' ' is pressed, meaning it moves back near the mouth of the gun.