Border, Gun, Bullet, Food Positioning
Added a border. Now that objects on the board are on coords that are a
multiple of 10, gun had to have its x-coord changed. There was also a
problem with the creation of the bullet from the gun which also had to
be fixed. Food pellet's random coordinate also shouldn't fall on the
boundary anymore.